Changed yaml format to accomodate objects better
Now the roguelike map that goes to the yaml file is not
a map of terrains but a map of "cells", where each character
represents a terrain type and a list of contents for the cell.
So for example a character might be represented by '@' in the
map, where the file would explain that @ meant a dungeon floor
plus a character and anything they were standing on.